# Fix orbital rotation failure due to PooledMemory.