--- a/src/QMCWaveFunctions/Fermion/SlaterDetOpt.h
+++ b/src/QMCWaveFunctions/Fermion/SlaterDetOpt.h
@@ -181,11 +181,11 @@ class SlaterDetOpt : public DiracDeterminantBase {
 
     void restore(int iat);
 
-    RealType registerData(ParticleSet& P, BufferType& buf);
+    void registerData(ParticleSet& P, WFBufferType& buf);
 
-    RealType updateBuffer(ParticleSet& P, BufferType& buf, bool fromscratch=false);
+    RealType updateBuffer(ParticleSet& P, WFBufferType& buf, bool fromscratch=false);
 
-    void copyFromBuffer(ParticleSet& P, BufferType& buf);
+    void copyFromBuffer(ParticleSet& P, WFBufferType& buf);
 
     OrbitalBasePtr makeClone(ParticleSet& tqp) const;